+++ CHANGELOG.md
@@ -12,6 +12,7 @@ Changelog
     ```
 - **Enhancement**: Improve IR caching layer with supertype caching.
 - **Fix**: Don't treat `l` as an illegal char in name allocating. This was supposed to be `;`.
+- **Fix**: Ensure all reserved properties by child graphs get properties in parents. Previously, there were some cases for bindings that were unused in the parent and otherwise did not meet the criteria for having a backing property would get missed in binding property collection.
 - Split `javax` and `jakarta` interop APIs into separate artifacts from dagger for reuse.
 
 0.7.4